// types.ts
export interface JourneyItem {
  companyKey: string;
  roleKey: string;
  period: string;
  logoUrl: string;
}

export interface SocialLink {
  key: string;
  name: string;
  icon: string;
  url: string;
  color: string;
}

export interface EcosystemProject {
  nameKey: string;
  descriptionKey: string;
}